% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_html.R
\name{convert_entrez_to_symbols}
\alias{convert_entrez_to_symbols}
\title{Convert ENTREZ IDs to gene symbols}
\usage{
convert_entrez_to_symbols(entrez_ids, organism)
}
\description{
Convert ENTREZ IDs to gene symbols
}
\keyword{internal}
