% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_html.R
\name{generate_results_table_html}
\alias{generate_results_table_html}
\title{Generate HTML table for all results}
\usage{
generate_results_table_html(result_df)
}
\description{
Generate HTML table for all results
}
\keyword{internal}
