% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_html.R
\name{generate_top_pathways_html}
\alias{generate_top_pathways_html}
\title{Generate HTML for top pathways section}
\usage{
generate_top_pathways_html(result_df, n = 5, organism = "human")
}
\description{
Generate HTML for top pathways section
}
\keyword{internal}
