% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot GSEA Results}
\usage{
plot_gsea(x, type = "dotplot", top_n = 20, ...)
}
\arguments{
\item{x}{A gseaResult object from gsea_from_deseq()}

\item{type}{Plot type: "dotplot" (default), "barplot", or "enrichment"}

\item{top_n}{Number of top pathways to show (default: 20)}

\item{...}{Additional arguments passed to clusterProfiler plotting functions}
}
\value{
A ggplot2 object
}
\description{
Create publication-ready plots from GSEA results.
}
\examples{
\dontrun{
results <- gsea_from_deseq(dds, contrast = c("condition", "treated", "control"))
plot_gsea(results)
plot_gsea(results, type = "barplot", top_n = 15)
}

}
